feat: Add environment ID support for hooks. (#81)
This PR pipes the support for the environment ID header from the event
source, through the data sources, and into the data store, and then
exposes it to the evaluation series hook context.

Each step can optionally support the environment ID/headers for
compatibility.

This PR will be held for back-end support and testing.