feat: add all available attributes to a span (#167)
**Requirements**

- [x] I have added test coverage for new or changed functionality
- [x] I have followed the repository's [pull request submission
guidelines](../blob/main/CONTRIBUTING.md#submitting-pull-requests)
- [x] I have validated my changes against all supported platform
versions

**Related issues**

N/A: I can create one if necessary

**Describe the solution you've provided**

This PR adds the ability to capture all available attributes when in
"span mode". It will only add the additional attributes when
`IncludeAllAttributesInSpan` is `true`; otherwise, the existing behavior
is retained.

This change is useful because it's easier to work with spans than span
events in some observability tools. With this change, spans can have the
same attributes as span events.

**Describe alternatives you've considered**

We could leave it as is but it would make spans less useful to
consumers.

**Additional context**

I talked to @kinyoklion about this on Slack about this change prior to
opening the PR.

Signed-off-by: Michael Beemer <beeme1mr@users.noreply.github.com>