chore: Prevent race condition in DynamoDB Tests
DynamoDB tests share a database / table and reset it for each test. XUnit tests can run in parallel causing one test to reset the table in the middle of another test executing and triggering a failure. This can be observed by adding a delay after a write and before a read.

Collections will ensure the specific tests are not run in parallel. I tested using a unique prefix for each test but still observed the error. Ideally these should be able to run in parallel but this is a compromise of invested time to fix it vs more stable tests now.